Introduce transitions to go to search (#178)
This commit adds the logic to perform a shared element transition
between `MainActivity` and `SearchActivity`.
I've tried to implement this using the Android Navigation Component
that comes in Android Jetpack but I've found an issue: when navigating
back there's a weird blink that I couldn't get rid of. That's why I decided
to implement this without the Navigation Component and that's where 
`MainToSearchNavigationDelegate` comes in.